pyproject.toml
[project]
name = "clonebot"
version = "0.1.0"
description = "Digital person clone - chat with memories in their voice and style"
readme = "README.md"
authors = [
    { name = "maxmcoste", email = "maxmcold@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "ollama>=0.3.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.0.0",
]

[project.scripts]
clonebot = "clonebot.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clonebot"]
